chore: fix circular dependency issue
